+++ com.peopleground.sagwim.fe/src/components/common/AlertDialog.tsx
@@ -1,5 +1,5 @@
-import { useEffect } from 'react'
 import { createPortal } from 'react-dom'
+import { useModalKeyboard } from '../../hooks/useModalKeyboard'
 import styles from './AlertDialog.module.css'
 
 type AlertVariant = 'success' | 'error'
@@ -19,14 +19,7 @@ export function AlertDialog({
   onClose,
   confirmLabel = '확인',
 }: AlertDialogProps) {
-  useEffect(() => {
-    if (!isOpen) return
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' || e.key === 'Enter') onClose()
-    }
-    document.addEventListener('keydown', handleKeyDown)
-    return () => document.removeEventListener('keydown', handleKeyDown)
-  }, [isOpen, onClose])
+  useModalKeyboard(isOpen, onClose, ['Escape', 'Enter'])
 
   if (!isOpen) return null
