+++ com.peopleground.sagwim.fe/src/components/common/SuccessDialog.tsx
@@ -1,5 +1,6 @@
 import { useEffect } from 'react'
 import { createPortal } from 'react-dom'
+import { useModalKeyboard } from '../../hooks/useModalKeyboard'
 import styles from './SuccessDialog.module.css'
 
 interface SuccessDialogProps {
@@ -23,14 +24,7 @@ export function SuccessDialog({
   onClose,
   confirmLabel = '확인',
 }: SuccessDialogProps) {
-  useEffect(() => {
-    if (!isOpen) return
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' || e.key === 'Enter') onClose()
-    }
-    document.addEventListener('keydown', handleKeyDown)
-    return () => document.removeEventListener('keydown', handleKeyDown)
-  }, [isOpen, onClose])
+  useModalKeyboard(isOpen, onClose, ['Escape', 'Enter'])
 
   useEffect(() => {
     if (!isOpen) return
